+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changes in 0.10.3 (2019-xx-xx)
 Improvements:
  * Integrations: Use the integrations manager provided by the homeserver admin via .well-known (#2815).
  * i18n: Add Welsh (cy).
+ * SerializationService: Add deserialisation of Any.
 
 Changes in 0.10.2 (2019-11-15)
 ===============================================